introduced clerk into the application to speed run the authentication
