+++ src/app/login/[[...login]]/page.tsx
@@ -1,5 +1,5 @@
 import React from "react";
-import Login from "../components/Login";
+import Login from "../../components/Login";
 
 export default function Page() {
   return (
