+++ src/app/page.tsx
@@ -1,10 +1,374 @@
 import Link from "next/link";
+import Image from "next/image";
 import { Hero } from "./components/Hero";
+import VoxyLogo from "@/public/V-removebg-preview.png";
+import { SignUpHero } from "./components/Signup/SIgnUpHero";
 
 export default function Home() {
   return (
     <main className="flex justify-center text-center flex-col min-h-screen">
-      <Hero />
+      <header className="">
+        <div className="mx-auto flex h-16 max-w-screen-xl items-center gap-8 px-4 sm:px-6 lg:px-8">
+          <Link className="block text-[#DB4A2B]" href="/">
+            <span className="sr-only">Home</span>
+            <Image src={VoxyLogo} alt="Voxy Logo" width={100} height={100} />
+          </Link>
+
+          <div className="flex flex-1 justify-end">
+            <div className="flex items-center gap-4">
+              <div className="sm:flex sm:gap-4">
+                <Link
+                  className="block rounded-md bg-[#DB4A2B] px-5 py-2.5 text-sm font-medium text-white transition hover:opacity-75"
+                  href="/login"
+                >
+                  Login
+                </Link>
+                <Link
+                  className="hidden rounded-md bg-gray-100 px-5 py-2.5 text-sm font-medium text-[#DB4A2B] transition hover:opacity sm:block"
+                  href="/signup"
+                >
+                  Register
+                </Link>
+              </div>
+            </div>
+          </div>
+        </div>
+      </header>
+      <main>
+        <SignUpHero />
+        <div className="text-center">
+          <h1 className="text-4xl font-semibold text-white">
+            We are the future of <br />
+            <span className="text-4xl md:text-[6rem] font-bold mt-1 leading-none">
+              Training with Voice AI.
+            </span>
+          </h1>
+        </div>
+
+        <div className="p-10 md:p-20 flex flex-wrap mt-10 justify-center gap-20 items-center">
+          <div className="relative flex w-80 flex-col rounded-xl bg-white bg-clip-border text-gray-700 shadow-md">
+            <div className="relative mx-4 -mt-6 h-40 overflow-hidden rounded-xl bg-blue-gray-500 bg-clip-border text-white shadow-lg shadow-blue-gray-500/40 bg-gradient-to-r from-[#51ba69] to-[#51ba69]"></div>
+            <div className="p-6">
+              <h5 className="mb-2 block font-sans text-xl font-semibold leading-snug tracking-normal text-blue-gray-900 antialiased">
+                Advanced Feedback
+              </h5>
+              <p className="block font-sans text-base font-light leading-relaxed text-inherit antialiased">
+                We provide advanced feedback on your training sessions and
+                explain how you can improve.
+              </p>
+            </div>
+            <div className="p-6 pt-0">
+              <button
+                data-ripple-light="true"
+                type="button"
+                className="select-none rounded-lg bg-[#51ba69] py-3 px-6 text-center align-middle font-sans text-xs font-bold uppercase text-white shadow-md shadow-blue-500/20 transition-all hover:shadow-lg hover:shadow-[#51ba69]-500/40 focus:opacity-[0.85] focus:shadow-none active:opacity-[0.85] active:shadow-none disabled:pointer-events-none disabled:opacity-50 disabled:shadow-none"
+              >
+                Read More
+              </button>
+            </div>
+          </div>
+          <div className="relative flex w-80 flex-col rounded-xl bg-white bg-clip-border text-gray-700 shadow-md">
+            <div className="relative mx-4 -mt-6 h-40 overflow-hidden rounded-xl bg-blue-gray-500 bg-clip-border text-white shadow-lg shadow-blue-gray-500/40 bg-gradient-to-r from-[#8e55ea] to-[#8e55ea]"></div>
+            <div className="p-6">
+              <h5 className="mb-2 block font-sans text-xl font-semibold leading-snug tracking-normal text-blue-gray-900 antialiased">
+                Addaptive Learning
+              </h5>
+              <p className="block font-sans text-base font-light leading-relaxed text-inherit antialiased">
+                Our AI spots your weakness and strengths and adapts the training
+                to suit your needs.
+              </p>
+            </div>
+            <div className="p-6 pt-0">
+              <button
+                data-ripple-light="true"
+                type="button"
+                className="select-none rounded-lg bg-[#8e55ea] py-3 px-6 text-center align-middle font-sans text-xs font-bold uppercase text-white shadow-md shadow-blue-500/20 transition-all hover:shadow-lg hover:shadow-[#8e55ea]-500/40 focus:opacity-[0.85] focus:shadow-none active:opacity-[0.85] active:shadow-none disabled:pointer-events-none disabled:opacity-50 disabled:shadow-none"
+              >
+                Read More
+              </button>
+            </div>
+          </div>
+          <div className="relative flex w-80 flex-col rounded-xl bg-white bg-clip-border text-gray-700 shadow-md">
+            <div className="relative mx-4 -mt-6 h-40 overflow-hidden rounded-xl bg-blue-gray-500 bg-clip-border text-white shadow-lg shadow-blue-gray-500/40 bg-gradient-to-r from-blue-500 to-blue-600"></div>
+            <div className="p-6">
+              <h5 className="mb-2 block font-sans text-xl font-semibold leading-snug tracking-normal text-blue-gray-900 antialiased">
+                Session Transcription
+              </h5>
+              <p className="block font-sans text-base font-light leading-relaxed text-inherit antialiased">
+                Our platform transcribes your training sessions and provides you
+                with a detailed report.
+              </p>
+            </div>
+            <div className="p-6 pt-0">
+              <button
+                data-ripple-light="true"
+                type="button"
+                className="select-none rounded-lg bg-blue-500 py-3 px-6 text-center align-middle font-sans text-xs font-bold uppercase text-white shadow-md shadow-blue-500/20 transition-all hover:shadow-lg hover:shadow-blue-500/40 focus:opacity-[0.85] focus:shadow-none active:opacity-[0.85] active:shadow-none disabled:pointer-events-none disabled:opacity-50 disabled:shadow-none"
+              >
+                Read More
+              </button>
+            </div>
+          </div>
+        </div>
+
+        <div className="text-center p-10 md:p-20">
+          <h1 className="text-4xl font-semibold text-white">
+            Start your journey <br />
+            <span className="text-4xl md:text-[6rem] font-bold mt-1 leading-none">
+              Our Pricing.
+            </span>
+          </h1>
+        </div>
+
+        <div className="mx-auto max-w-3xl px-4 py-8 sm:px-6 sm:py-12 lg:px-8">
+          <div className="grid grid-cols-1 gap-4 sm:grid-cols-2 sm:items-center md:gap-8">
+            <div className="rounded-2xl border border-indigo-600 p-6 shadow-sm ring-1 ring-indigo-600 sm:order-last sm:px-8 lg:p-12">
+              <div className="text-center">
+                <h2 className="text-lg font-medium text-white">
+                  Pro
+                  <span className="sr-only">Plan</span>
+                </h2>
+
+                <p className="mt-2 sm:mt-4">
+                  <strong className="text-3xl font-bold text-white sm:text-4xl">
+                    {" "}
+                    15${" "}
+                  </strong>
+
+                  <span className="text-sm font-medium text-white">/month</span>
+                </p>
+              </div>
+
+              <ul className="mt-6 space-y-2">
+                <li className="flex items-center gap-1">
+                  <svg
+                    xmlns="http://www.w3.org/2000/svg"
+                    fill="none"
+                    viewBox="0 0 24 24"
+                    strokeWidth="1.5"
+                    stroke="currentColor"
+                    className="size-5 text-indigo-700"
+                  >
+                    <path
+                      strokeLinecap="round"
+                      strokeLinejoin="round"
+                      d="M4.5 12.75l6 6 9-13.5"
+                    />
+                  </svg>
+
+                  <span className="text-white">
+                    {" "}
+                    <span className="font-bold">UNLIMITED</span> sessions{" "}
+                  </span>
+                </li>
+
+                <li className="flex items-center gap-1">
+                  <svg
+                    xmlns="http://www.w3.org/2000/svg"
+                    fill="none"
+                    viewBox="0 0 24 24"
+                    strokeWidth="1.5"
+                    stroke="currentColor"
+                    className="size-5 text-indigo-700"
+                  >
+                    <path
+                      strokeLinecap="round"
+                      strokeLinejoin="round"
+                      d="M4.5 12.75l6 6 9-13.5"
+                    />
+                  </svg>
+
+                  <span className="text-white"> Session Feedback </span>
+                </li>
+
+                <li className="flex items-center gap-1">
+                  <svg
+                    xmlns="http://www.w3.org/2000/svg"
+                    fill="none"
+                    viewBox="0 0 24 24"
+                    strokeWidth="1.5"
+                    stroke="currentColor"
+                    className="size-5 text-indigo-700"
+                  >
+                    <path
+                      strokeLinecap="round"
+                      strokeLinejoin="round"
+                      d="M4.5 12.75l6 6 9-13.5"
+                    />
+                  </svg>
+
+                  <span className="text-white"> Session Transcription </span>
+                </li>
+
+                {/* <li className="flex items-center gap-1">
+                  <svg
+                    xmlns="http://www.w3.org/2000/svg"
+                    fill="none"
+                    viewBox="0 0 24 24"
+                    strokeWidth="1.5"
+                    stroke="currentColor"
+                    className="size-5 text-indigo-700"
+                  >
+                    <path
+                      strokeLinecap="round"
+                      strokeLinejoin="round"
+                      d="M4.5 12.75l6 6 9-13.5"
+                    />
+                  </svg>
+
+                  <span className="text-white"> Help center access </span>
+                </li>
+
+                <li className="flex items-center gap-1">
+                  <svg
+                    xmlns="http://www.w3.org/2000/svg"
+                    fill="none"
+                    viewBox="0 0 24 24"
+                    strokeWidth="1.5"
+                    stroke="currentColor"
+                    className="size-5 text-indigo-700"
+                  >
+                    <path
+                      strokeLinecap="round"
+                      strokeLinejoin="round"
+                      d="M4.5 12.75l6 6 9-13.5"
+                    />
+                  </svg>
+
+                  <span className="text-white"> Phone support </span>
+                </li>
+
+                <li className="flex items-center gap-1">
+                  <svg
+                    xmlns="http://www.w3.org/2000/svg"
+                    fill="none"
+                    viewBox="0 0 24 24"
+                    strokeWidth="1.5"
+                    stroke="currentColor"
+                    className="size-5 text-indigo-700"
+                  >
+                    <path
+                      strokeLinecap="round"
+                      strokeLinejoin="round"
+                      d="M4.5 12.75l6 6 9-13.5"
+                    />
+                  </svg>
+
+                  <span className="text-white"> Community access </span>
+                </li> */}
+              </ul>
+
+              <Link
+                href="/signup/form"
+                className="mt-8 block rounded-full border border-indigo-600 bg-indigo-600 px-12 py-3 text-center text-sm font-medium text-white hover:bg-indigo-700 hover:ring-1 hover:ring-indigo-700 focus:outline-none focus:ring active:text-indigo-500"
+              >
+                Get Started
+              </Link>
+            </div>
+
+            <div className="rounded-2xl border border-gray-200 p-6 shadow-sm sm:px-8 lg:p-12">
+              <div className="text-center">
+                <h2 className="text-lg font-medium text-white">
+                  Starter
+                  <span className="sr-only">Plan</span>
+                </h2>
+
+                <p className="mt-2 sm:mt-4">
+                  <strong className="text-3xl font-bold text-white sm:text-4xl">
+                    {" "}
+                    Free{" "}
+                  </strong>
+                </p>
+              </div>
+
+              <ul className="mt-6 space-y-2">
+                <li className="flex items-center gap-1">
+                  <svg
+                    xmlns="http://www.w3.org/2000/svg"
+                    fill="none"
+                    viewBox="0 0 24 24"
+                    strokeWidth="1.5"
+                    stroke="currentColor"
+                    className="size-5 text-indigo-700"
+                  >
+                    <path
+                      strokeLinecap="round"
+                      strokeLinejoin="round"
+                      d="M4.5 12.75l6 6 9-13.5"
+                    />
+                  </svg>
+
+                  <span className="text-white"> 10 Sessions a month </span>
+                </li>
+                {/* 
+                <li className="flex items-center gap-1">
+                  <svg
+                    xmlns="http://www.w3.org/2000/svg"
+                    fill="none"
+                    viewBox="0 0 24 24"
+                    strokeWidth="1.5"
+                    stroke="currentColor"
+                    className="size-5 text-indigo-700"
+                  >
+                    <path
+                      strokeLinecap="round"
+                      strokeLinejoin="round"
+                      d="M4.5 12.75l6 6 9-13.5"
+                    />
+                  </svg>
+
+                  <span className="text-white"> 2GB of storage </span>
+                </li>
+
+                <li className="flex items-center gap-1">
+                  <svg
+                    xmlns="http://www.w3.org/2000/svg"
+                    fill="none"
+                    viewBox="0 0 24 24"
+                    strokeWidth="1.5"
+                    stroke="currentColor"
+                    className="size-5 text-indigo-700"
+                  >
+                    <path
+                      strokeLinecap="round"
+                      strokeLinejoin="round"
+                      d="M4.5 12.75l6 6 9-13.5"
+                    />
+                  </svg>
+
+                  <span className="text-white"> Email support </span>
+                </li>
+
+                <li className="flex items-center gap-1">
+                  <svg
+                    xmlns="http://www.w3.org/2000/svg"
+                    fill="none"
+                    viewBox="0 0 24 24"
+                    strokeWidth="1.5"
+                    stroke="currentColor"
+                    className="size-5 text-indigo-700"
+                  >
+                    <path
+                      strokeLinecap="round"
+                      strokeLinejoin="round"
+                      d="M4.5 12.75l6 6 9-13.5"
+                    />
+                  </svg>
+
+                  <span className="text-white"> Help center access </span>
+                </li> */}
+              </ul>
+
+              <Link
+                href="/signup/form"
+                className="mt-8 block rounded-full border border-indigo-600 bg-white px-12 py-3 text-center text-sm font-medium text-indigo-600 hover:ring-1 hover:ring-indigo-600 focus:outline-none focus:ring active:text-indigo-500"
+              >
+                Get Started
+              </Link>
+            </div>
+          </div>
+        </div>
+      </main>
     </main>
   );
 }
