+++ src/app/signup/components/Step.tsx
